ACL_PyTorch/built-in/nlp/Pet_for_Pytorch/requirements.txt
paddlepaddle==2.3.1
paddlenlp==2.3.4
paddlefsl==1.1.0
paddle2onnx==0.9.8
onnx==1.12.0
onnxruntime==1.11.1
numpy==1.21.6
datasets==2.3.2
torch==1.10.1
tqdm==4.63.1
